@@ -1,19 +1,18 @@
 [package]
-name = "skedge"
-description = "Ergonomic single-process job scheduling for Rust programs."
-version = "0.1.2"
-edition = "2018"
 authors = ["Ben Lovy <ben@deciduously.com>"]
+description = "Ergonomic single-process job scheduling for Rust programs."
 documentation = "https://docs.rs/skedge"
+edition = "2021"
 homepage = "https://crates.io/crates/skedge"
 include = ["**/*.rs", "Cargo.toml"]
 keywords = ["utility", "scheduling"]
 license = "BSD-3-Clause"
+name = "skedge"
 readme = "README.md"
 repository = "https://github.com/deciduously/skedge"
+version = "0.1.3"
 
 [lib]
-
 crate-type = ["rlib", "cdylib"]
 
 [features]
@@ -21,20 +20,18 @@ default = []
 ffi = []
 
 [dependencies]
-
 chrono = "0.4"
-log = "0.4"
 pretty_env_logger = "0.4"
 lazy_static = "1.4"
-libc = { version = "0.2.0" }
+libc = { version = "0.2" }
 rand = "0.8"
 regex = "1.5"
-thiserror = "1"
+thiserror = "1.0"
+tracing = "0.1"
 
 [dev-dependencies]
+pretty_assertions = "1.2"
 
-pretty_assertions = "0.7"
-
-[profile]
 [profile.release]
 lto = true
+strip = true
